New useContactLink for SettingsPage
